Made Major Changes...
Alpha version of the game has now started. I got the basic menu functionality started and am still figuring out how to get other things working. 

I also reconfigured the image file names to the standard Java notation. This means that there are also changes to the directories of the per-alpha-test files to account for that change. Please take the time to get the changes added to your workspace so the naming conventions are followed for the rest of the project.