Updated READMEs...
Updated the package READMEs to specify the rules around making Java file backups of previous versions worked on. This allows for us to have records of older work in case any changes wished to be reverted - we can see what was done and keep track of the changes. 

Please view the respective package READMEs for more info on how to properly set up backup folders...

Additional Changes:
- New variable established (File file) to try and test loading other Java programs; not in working order at the moment... (06/04/2022) [src/alpha_versions/0.0.1/HaevinChess.java]