Moving into Beta...
Major changes have been made to the project thus far. 

Changes are as follows:
Implented 90% of the pre_alpha_test work done by Alvin and Hammad. The program now loads the game board, images, as well as some other behaviour. Piece controls and movements are to be added (06/12/2022). 

Note: We will need to have all piece movement/behaviours done by tuesday to be on time for the rest of the project!