Moving into Beta...
Please use the beta_versions package now, as we have moved onto the beta program. View the inline comments left in the code to gain an understanding of what has been done and why.

Additional Changes:
Got all piece movements and behaviours complete. A checkKingScanner() and checkGameScanner() methods will need to be created as well. (06/14/2022)