Major Changes Made...
ALL PROJECTS NEED TO BE REDOWNLOADED!

The project is now nearing completion in most areas. The changes made below should make it easier to manage the other things that must be completed...

Patch Notes:
- Broke down the entire game into multiple class files. Each one contains a different aspect of the game and its properties. Not fully finalized but will be the new way in which we will manage variables and other things in the game to prevent the main class file (this one!!) from being too large and hard to manage (06/18/2022 -> Evan)
- All piece properties and condition behaviours are now stored in their respective "~Properties.java" file (06/18/2022 -> Evan)
- Added restart functionality to the game (06/18/2022 -> Evan)
- Implemented any code that was added in yesterday (06/18/2022 -> Evan)

To now load the game, it's as simple as last time. Have any of the classes open and compile/run. Since the main is only located in "HaevinChess.java", as long as the launch is taking place in the same root directory as that class file, the entire program should load just fine. Please let me know if any issues occur!

Note: "~Properties.java" files contain all the code within methods called in the main class (HaevinChess.java). DO NOT IMPLEMENT A STATIC MAIN METHOD IN ANY JAVA CLASSES OUTSIDE OF THE MAIN PROGRAM! This may cause compiling errors! If any new piece-specific rules/moves are to be created, code them in their respective "~Properties.java" files. If things are to be added that do not correlate to any of the existing methods, please create a new one!

Fun Fact: Due to all these changes meant to simplify the program, "HaevinChess.java" went from 1300 lines of code to exactly 643! WTF